% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savvy-package.R
\docType{package}
\name{savvy-package}
\alias{savvy}
\alias{savvy-package}
\title{A Simple Wrapper of 'savvy-cli' Command}
\description{
Generates the C and R wrappers for the Rust-powered R package using the savvy framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yutannihilation/savvy}
}

}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
