% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{savvy_extract_tests}
\alias{savvy_extract_tests}
\title{Execute \verb{savvy-cli extract-tests}}
\usage{
savvy_extract_tests(path = "./src/rust/")
}
\arguments{
\item{path}{Path to the root of a Rust crate.}
}
\description{
Execute \verb{savvy-cli extract-tests}
}
