% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{savvy_source}
\alias{savvy_source}
\title{Compile Rust Code And Load}
\usage{
savvy_source(
  code,
  use_cache_dir = TRUE,
  env = parent.frame(),
  dependencies = list(),
  clean = NULL
)
}
\arguments{
\item{code}{Rust code to compile.}

\item{use_cache_dir}{If \code{TRUE}, reuse and override the cache dir to avoid
re-compilation. This is an expert-only option.}

\item{env}{The R environment where the R wrapping functions should be defined.}

\item{dependencies}{List of dependencies. (e.g. \code{list(once_cell = list(version = "1"))})
at the end of the R session.}

\item{clean}{If \code{TRUE}, remove the temporary R package used for compilation}
}
\description{
Compile Rust Code And Load
}
