% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{update_savvy_cli}
\alias{update_savvy_cli}
\title{Update 'savvy-cli'}
\usage{
update_savvy_cli(force = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, download the savvy CLI even if it's up-to-date.}
}
\description{
Update 'savvy-cli'
}
