% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{savvy_source}
\alias{savvy_source}
\title{Compile Rust Code And Load}
\usage{
savvy_source(code, use_cache_dir = TRUE)
}
\arguments{
\item{code}{Rust code to compile.}

\item{use_cache_dir}{If \code{TRUE}, reuse and override the cache dir to avoid re-compilation.}
}
\description{
Compile Rust Code And Load
}
