% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savvy_update}
\alias{savvy_update}
\title{Execute `savvy-cli update``}
\usage{
savvy_update(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{Path to the root of an R package}

\item{verbose}{If \code{TRUE}, show all the output from savvy-cli.}
}
\description{
Execute `savvy-cli update``
}
